% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_PELMO.R
\name{install_PELMO}
\alias{install_PELMO}
\title{Download FOCUS PELMO from the official site and unzip into the package installation directory}
\usage{
install_PELMO(version = "5.5.3", local_zip = NULL)
}
\arguments{
\item{version}{The FOCUS PELMO version to install}
}
\description{
Download FOCUS PELMO from the official site and unzip into the package installation directory
}
